fix(add_project): Replaced auth token with session
