+++ src/components/ProjectAddSummary/component/LinkedReleasesAndProjects/LinkedReleases.tsx
@@ -9,41 +9,41 @@
 
 'use client'
 
-import { Table } from '@/components/sw360'
+// import { Table } from '@/components/sw360'
 import LinkedReleasesModal from '@/components/sw360/LinkedReleasesModal/LinkedReleasesModal'
 import { useState } from 'react'
 
 export default function LinkedReleases() {
     const [showLinkedReleasesModal, setShowLinkedReleasesModal] = useState(false)
-    const [linkedRelease, setLinkedRelease] = useState([])
+    // const [linkedRelease, setLinkedRelease] = useState([])
 
-    const columns = [
-        {
-            id: 'name',
-            name: 'Release Name',
-            sort: true,
-        },
-        {
-            id: 'version',
-            name: 'Release Version',
-            sort: true,
-        },
-        {
-            id: 'relation',
-            name: 'Release Relation',
-            sort: true,
-        },
-        {
-            id: 'mainlineState',
-            name: 'Project Mainline State',
-            sort: true,
-        },
-        {
-            id: 'comments',
-            name: 'Comments',
-            sort: true,
-        }
-    ]
+    // const columns = [
+    //     {
+    //         id: 'name',
+    //         name: 'Release Name',
+    //         sort: true,
+    //     },
+    //     {
+    //         id: 'version',
+    //         name: 'Release Version',
+    //         sort: true,
+    //     },
+    //     {
+    //         id: 'relation',
+    //         name: 'Release Relation',
+    //         sort: true,
+    //     },
+    //     {
+    //         id: 'mainlineState',
+    //         name: 'Project Mainline State',
+    //         sort: true,
+    //     },
+    //     {
+    //         id: 'comments',
+    //         name: 'Comments',
+    //         sort: true,
+    //     }
+    // ]
 
     return (
         <>
@@ -55,9 +55,7 @@ export default function LinkedReleases() {
                         <hr className='my-2 mb-2' style={{ color: '#5D8EA9', paddingLeft: '0px' }} />
                     </h6>
                 </div>
-                <div style={{paddingLeft: '0px' }}>
-                    <Table data={linkedRelease} columns={columns}/>
-                </div>
+                <div style={{ paddingLeft: '0px' }}>{/* <Table data={linkedRelease} columns={columns}/> */}</div>
                 <div className='row' style={{ paddingLeft: '0px' }}>
                     <div className='col-lg-4'>
                         <button
